Fixed a bug in bulk allocation
The bulk allocation of students to projects didn't use the same path
as regular staff accepting a student, so wasn't triggering the errors
if a project was full.

Now re-jigged so it checks and bails out as soon as something has
gone wrong.

Possibly re-do as a vuejs form so it can flag up the error in real
time & before submission...