Added some auth tests and fixed weird error with student project list
The student project list was _sometimes_ not working right.  It was
coming from laravel as an 'object of objects' rather than an 'array
of objects'.  Something to do with the collection()->toJson() mapping
I guess.

Why it wasn't breaking the JS side of things every time is a bit of
a mystery.  Now changed to use ye olde foreach php array building
which seems (ha) to work ok.