Changed the view that students get on login, deleted old cruft
Was asked to change what happens when a student logs in.  Now if they
are already accepted onto a project they only get to see that.  If they
have made project choices they just get to see that list and can no longer
change their minds.

Also deleted a lot of cruft that was left over from the first 'test'
version of the code.